% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drg_statistics.R
\name{drg_statistics}
\alias{drg_statistics}
\title{Statistics for DRG}
\usage{
drg_statistics(option)
}
\arguments{
\item{option}{a string name for statistics mean, median, or standard deviation}
}
\value{
A statistical calculation for mean or median or standard deviation
}
\description{
This function produces a table of either mean, median, or standard deviations
of all the 3 digit DRG codes for average Medical payments.
}
\examples{
drg_statistics("mean")
}
