% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drg_to_payment.R
\name{drg_to_payment}
\alias{drg_to_payment}
\title{Boxplot of payments by DRG code}
\usage{
drg_to_payment(option)
}
\arguments{
\item{option}{}
}
\value{

}
\description{
Boxplot of payments by DRG code
}
