% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drg_analysis.R
\name{get_prevalence_rates}
\alias{get_prevalence_rates}
\title{Get prevalence rates for all disease categories}
\usage{
get_prevalence_rates(conn, cdm_schema)
}
\arguments{
\item{conn}{Database connection}

\item{cdm_schema}{CDM schema name}
}
\value{
Data frame with counts and prevalence rates
}
\description{
Get prevalence rates for all disease categories
}
