% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drg_analysis.R
\name{get_total_patients}
\alias{get_total_patients}
\title{Get total number of patients in database}
\usage{
get_total_patients(conn, cdm_schema)
}
\arguments{
\item{conn}{Database connection}

\item{cdm_schema}{CDM schema name}
}
\value{
Total number of patients
}
\description{
Get total number of patients in database
}
