% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicare_stat.R
\name{medicare_stat}
\alias{medicare_stat}
\title{Title}
\usage{
medicare_stat(df, funx)
}
\arguments{
\item{df}{a dataframe}

\item{funx}{a function of statistic input}
}
\value{
A table with a table that provides the result of funx
}
\description{
Title
}
\examples{
Medicare_stat (DRG, mean)


}
