% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payment_plot.R
\name{payment_plot}
\alias{payment_plot}
\title{Title}
\usage{
payment_plot(df, x)
}
\arguments{
\item{df}{a dataframe}

\item{x}{a column name with type of payment from the df}
}
\value{
A boxplot with DRG code at x-axis and type of payment at y-axis, colored by various DRG code
}
\description{
Title
}
\examples{

DRG_payment_plot (DRG, 'Average Total Payments')

}
