% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drg_to_payment.R
\name{drg_to_payment}
\alias{drg_to_payment}
\title{Boxplot for DRG}
\usage{
drg_to_payment(option)
}
\arguments{
\item{option}{a string name for different payments: Medicare, Total, or Covered.}
}
\value{
A boxplot for the average Medicare payments, the average total payment, or the average covered charges
}
\description{
This function produces a boxplot of payments by DRG code.
}
\examples{
drg_statistics("mean")
}
